requirements.txt
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
discord-protos==0.0.2
discord.py-self==2.0.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
protobuf==5.27.0
PyNaCl==1.4.0
Werkzeug==3.0.3
yarl==1.9.4
